+++ src/stats_logger.hpp
@@ -16,8 +16,6 @@ private:
 
   std::atomic<uint64_t> m_active_clients{0};
   std::atomic<uint64_t> m_max_clients{0};
-  std::atomic<uint64_t> m_total_requests{0};
-  std::chrono::steady_clock::time_point m_start_time;
 
   std::jthread m_log_thread;
   std::condition_variable_any m_cv;
@@ -34,11 +32,6 @@ public:
 
   void increment_active_clients();
   void decrement_active_clients();
-  void increment_total_requests();
-
-  uint64_t get_active_clients() const { return m_active_clients.load(); }
-  uint64_t get_max_clients() const { return m_max_clients.load(); }
-  uint64_t get_total_requests() const { return m_total_requests.load(); }
 
   void start_periodic_logging();